No-op: gearing to optimize the priority queue so it can't have multiple copies of a stop in there
